Try adapt MVVM to MVVM with databinding I don't know why it doesn't work
