+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: 'kotlin-kapt'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.team4.BoulderBuild"
+        applicationId "com.team4.boulderBuild"
         minSdkVersion 22
         targetSdkVersion 28
         versionCode 1
@@ -18,6 +18,13 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
+    }
+    dataBinding {
+        enabled true
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 }
 
@@ -36,10 +43,15 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.0.0'
     implementation 'androidx.navigation:navigation-ui-ktx:2.0.0'
+    kapt 'com.android.databinding:compiler:3.2'
     kapt 'androidx.room:room-compiler:2.2.1'
     implementation 'androidx.room:room-runtime:2.2.1'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.2.2'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
+}
+
+kapt {
+    generateStubs = true
 }
