+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.team4.BoulderBuild">
+    package="com.team4.boulderBuild">
 
     <application
         android:allowBackup="true"
@@ -11,16 +11,16 @@
         android:supportsRtl="true"
         android:theme="@style/SplashTheme"
         tools:ignore="GoogleAppIndexingWarning"
-        android:name=".BoulderBuildApp">
+        android:name="com.team4.boulderBuild.BoulderBuildApp">
 
-        <activity android:name=".MainActivity">
+        <activity android:name="com.team4.boulderBuild.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.gymform.GymFormActivity">
+        <activity android:name="com.team4.boulderBuild.ui.gymform.GymFormActivity">
         </activity>
     </application>
