+++ app/src/main/java/com/team4/boulderBuild/BoulderBuildApp.kt
@@ -1,0 +1,13 @@
+package com.team4.boulderBuild
+
+import android.app.Application
+import com.team4.boulderBuild.config.ApplicationDI
+
+class BoulderBuildApp : Application(){
+
+    override fun onCreate() {
+        super.onCreate()
+
+        ApplicationDI.initDI(this)
+    }
+}
