+++ app/src/main/java/com/team4/boulderBuild/model/data/database/Gym.kt
@@ -1,0 +1,14 @@
+package com.team4.boulderBuild.model.data.database
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class Gym (
+    @PrimaryKey(autoGenerate = false) val id: Int,
+    val name: String,
+    val description: String,
+    val lat : Double,
+    val lon : Double,
+    val imgSrc : String
+)
