+++ app/src/main/java/com/team4/boulderBuild/ui/common/Scope.kt
@@ -1,0 +1,26 @@
+package com.team4.boulderBuild.ui.common
+
+import kotlin.coroutines.CoroutineContext
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
+
+interface Scope : CoroutineScope {
+
+    class Impl : Scope {
+        override lateinit var job: Job
+    }
+
+    var job: Job
+    override val coroutineContext: CoroutineContext
+        get() = Dispatchers.Main + job
+
+    fun initScope() {
+        job = SupervisorJob()
+    }
+
+    fun destroyScope() {
+        job.cancel()
+    }
+}
