+++ app/src/main/java/com/team4/boulderBuild/ui/common/ScopedViewModel.kt
@@ -1,0 +1,17 @@
+package com.team4.boulderBuild.ui.common
+
+import androidx.annotation.CallSuper
+import androidx.lifecycle.ViewModel
+
+abstract class ScopedViewModel : ViewModel(), Scope by Scope.Impl() {
+
+    init {
+        initScope()
+    }
+
+    @CallSuper
+    override fun onCleared() {
+        destroyScope()
+        super.onCleared()
+    }
+}
