+++ app/src/main/java/com/team4/boulderBuild/ui/dashboard/DashboardFragment.kt
@@ -1,0 +1,31 @@
+package com.team4.boulderBuild.ui.dashboard
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import com.team4.boulderBuild.R
+
+class DashboardFragment : Fragment() {
+
+    private lateinit var dashboardViewModel: DashboardViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        dashboardViewModel =
+            ViewModelProviders.of(this).get(DashboardViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_dashboard, container, false)
+        val textView: TextView = root.findViewById(R.id.text_dashboard)
+        dashboardViewModel.text.observe(this, Observer {
+            textView.text = it
+        })
+        return root
+    }
+}
