+++ app/src/main/java/com/team4/boulderBuild/ui/notifications/NotificationsFragment.kt
@@ -1,0 +1,31 @@
+package com.team4.boulderBuild.ui.notifications
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import com.team4.boulderBuild.R
+
+class NotificationsFragment : Fragment() {
+
+    private lateinit var notificationsViewModel: NotificationsViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        notificationsViewModel =
+            ViewModelProviders.of(this).get(NotificationsViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_notifications, container, false)
+        val textView: TextView = root.findViewById(R.id.text_notifications)
+        notificationsViewModel.text.observe(this, Observer {
+            textView.text = it
+        })
+        return root
+    }
+}
