+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_gyms"
-        android:name="com.team4.BoulderBuild.ui.gyms.GymsFragment"
+        android:name="com.team4.boulderBuild.ui.gyms.GymsFragment"
         android:label="@string/title_gyms"
         tools:layout="@layout/fragment_gyms" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.team4.BoulderBuild.ui.dashboard.DashboardFragment"
+        android:name="com.team4.boulderBuild.ui.dashboard.DashboardFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.team4.BoulderBuild.ui.notifications.NotificationsFragment"
+        android:name="com.team4.boulderBuild.ui.notifications.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
